+++ common/include/send.h
@@ -3,6 +3,6 @@
 
 #include "common.h"
 
-extern int send_request(int sockfd, int machine, uint32_t size, uint8_t path_len, uint8_t *path);
+extern int send_request(int sockfd, int machine, uint64_t size, uint8_t path_len, uint8_t *path);
 
 #endif /* end of include guard: NETRANS_SEND_H */
